Build fixes for imgui viewer
